move out common test functions
